+++ packages/svelte-integration/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Added the ability to hide logs
+- Added the ability to skip the plugin
+
 ## 1.0.1
 
 - Removed docs for editor documentation since it is not supported yet